Test appears to be suffering from changed stringification of floats
As part of today's Squashathon, this module was [marked as having a problem](https://github.com/perl6/ecosystem-unbitrot/issues/96).  Turns out it was a single test that was testing on the stringification of a float.  Since some work was done in that area recently, I assume that the changed stringification is the source of the problem, not different handling of floating point arithmetic itself.  Hence this change of the test.  Hope that's the right answer!